Interpret return statements
Since each statement is its own function as we interpret statements in a lox callable the functions are pushed on to the stack. Once we see a return statement we want to unwind up to the lox callable. this is kinda tricky so we use a hack and add a special error that holds the return value. then the logic that already exists to unwind on error is used to return a value out of the lox callable.